Remove bugs
動いたうおおおおお

+++ src/proc/Hand.cpp
@@ -6,7 +6,7 @@ void proc::Hand::begin(ssr2::Machine *machine) {
 
 void proc::Hand::update(ssr2::Machine *machine) {
     const ssr2::PS4Value &value = machine->currentPS4Value();
-    uint8_t angle = machine->hand.read();
+    int16_t angle = machine->hand.read();
     angle += value.l1 - value.r1;
     machine->hand.write(angle);
     #ifdef proc_verbose
